+++ Changelog.md
@@ -9,6 +9,7 @@ Features:
 
 Bugfixes:
  * Parser: Enforce commas between array and tuple elements.
+ * Parser: Limit maximum recursion depth.
 
 ### 0.4.15 (2017-08-08)